% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.dbgap.R
\name{browse.dbgap}
\alias{browse.dbgap}
\title{Opens the dbGaP webpage of your study in your browser}
\usage{
browse.dbgap(phs, no.browser = FALSE)
}
\arguments{
\item{phs}{dbGaP study ID (phs00xxxx, or 00xxxx, or xxx)}

\item{no.browser}{set on TRUE will just display the URL}
}
\value{
Opens the dbGaP webpage of your study in your browser
}
\description{
Opens the dbGaP webpage of your study in your browser
}
\author{
Gregoire Versmee, Laura Versmee, Mikael Dusenne, Niloofar Jalali
}
