% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.study.R
\name{browse.study}
\alias{browse.study}
\title{Opens the study webpage in a web browser}
\usage{
browse.study(phs, no.browser = FALSE)
}
\arguments{
\item{phs}{dbGaP study ID (phs00xxxx, or 00xxxx, or xxx)}

\item{no.browser}{set on TRUE will just display the URL}
}
\value{
Opens the study webpage in a web browser
}
\description{
Opens the study webpage in a web browser
}
\author{
Gregoire Versmee, Laura Versmee, Mikael Dusenne, Niloofar Jalali
}
