% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consent.groups.R
\name{consent.groups}
\alias{consent.groups}
\title{Gets the list of the study's consent groups}
\usage{
consent.groups(phs)
}
\arguments{
\item{phs}{dbGaP study ID (phs00xxxx, or 00xxxx, or xxx)}
}
\value{
Returns a data frame with 3 columns : "consent group number", "cg short name" and "cg long name"
}
\description{
Gets the list of the study's consent groups
}
\author{
Gregoire Versmee, Laura Versmee, Mikael Dusenne, Niloofar Jalali
}
