% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbgap.data_dict.R
\name{dbgap.data_dict}
\alias{dbgap.data_dict}
\title{Create a table dictionnary and a variable dictionnary}
\usage{
dbgap.data_dict(xml, dest)
}
\arguments{
\item{xml}{Folder where the xml files are located}

\item{dest}{Folder where the data dictionnary and the table dictionnary will be extracted}
}
\value{
2 csv files "phs_tables_dict.csv" and "phs_variables_dict.csv"
}
\description{
This function extracts informations from data.dict xml files in dbgap to create table dictionnary and a variable dictionnary.
}
\author{
Gregoire Versmee, Laura Versmee
}
