% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbgap.decrypt.R
\name{dbgap.decrypt}
\alias{dbgap.decrypt}
\title{Decrypts ncbi_enc files}
\usage{
dbgap.decrypt(files, key = FALSE)
}
\arguments{
\item{file}{file or folder where your encrypted files are located}
}
\value{
Decrypts the file(s) and replaces it in the same folder
}
\description{
This function decrypts dbGaP files (*ncbi_enc) using your personal key
}
\author{
Gregoire Versmee, Laura Versmee, Mikael Dusenne, Niloofar Jalali
}
