% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbgap.download.R
\name{dbgap.download}
\alias{dbgap.download}
\title{Downloads and decrypts dbGaP files}
\usage{
dbgap.download(krt, key = FALSE)
}
\arguments{
\item{krt}{path/to/your/cart.krt file generated from the dbGaP website}

\item{key}{path/to/yourkey.ngc file}
}
\value{
A folder named dbGaP-*** in your working directory that contains your decrypted files
}
\description{
This function downloads and decrypts dbGaP files
}
\author{
Gregoire Versmee, Laura Versmee, Mikael Dusenne, Niloofar Jalali
}
