% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.tables.R
\name{list.tables}
\alias{list.tables}
\title{Lists the phenotypic datatables in the study}
\usage{
list.tables(filelist)
}
\value{
Returns the phenotypic datatables in the study
}
\description{
Lists the phenotypic datatables in the study
}
\author{
Gregoire Versmee, Laura Versmee, Mikael Dusenne, Niloofar Jalali
}
