% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.variables.R
\name{list.variables}
\alias{list.variables}
\title{Gets the phenotypic variables in the study}
\usage{
list.variables(phs, useftp = F)
}
\arguments{
\item{phs}{dbGap study ID (phs00xxxx, or 00xxxx, or xxx)}
}
\value{
Return the phenotypic variables in the study
}
\description{
This function extracts informations from data.dict.xml files from the dbgap ftp server to get the study characteristics. Works only for a parent study.
}
\author{
Gregoire Versmee, Laura Versmee, Mikael Dusenne, Niloofar Jalali
}
