% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.tables.R
\name{n.tables}
\alias{n.tables}
\title{Gets the number of phenotypic datatables in the study}
\usage{
n.tables(phs)
}
\arguments{
\item{phs}{dbGap study ID (phs00xxxx, or 00xxxx, or xxx)}
}
\value{
Return the number of phenotypic datatables in the study
}
\description{
This function extracts informations from data.dict.xml files from the dbgap ftp server to get the study characteristics. Works only for a parent study.
}
\author{
Gregoire Versmee, Laura Versmee
}
