% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.variables.R
\name{n.variables}
\alias{n.variables}
\title{Gets the number of phenotypic variables in the study}
\usage{
n.variables(...)
}
\arguments{
\item{phs}{dbGaP study ID (phs00xxxx, or 00xxxx, or xxx)}
}
\value{
Returns the number of phenotypic variables in the study
}
\description{
This function extracts informations from data.dict.xml files from the dbGaP ftp server to get the study characteristics. Works only for a parent study.
}
\author{
Gregoire Versmee, Laura Versmee, Mikael Dusenne, Niloofar Jalali
}
