% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent.study.R
\name{parent.study}
\alias{parent.study}
\title{Gets you the parent study of the specified study}
\usage{
parent.study(phs)
}
\arguments{
\item{phs}{dbGap study ID (phs00xxxx, or 00xxxx, or xxx)}
}
\value{
Returns the phs id of the latest version of the parent study, and its full name. If the study is it's parent itself, will return the phs id and the name of the study.
}
\description{
Gets you the parent study of the specified study
}
\author{
Gregoire Versmee, Laura Versmee
}
