% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.dbgap.R
\name{search.dbgap}
\alias{search.dbgap}
\title{Gets the results of your search on dbGaP}
\usage{
search.dbgap(term, no.browser = FALSE)
}
\arguments{
\item{term}{Term of your search}

\item{no.browser}{set on TRUE will just display the URL}
}
\value{
A data frame of the results of your search on dbGaP
}
\description{
Gets the results of your search on dbGaP
}
\author{
Gregoire Versmee, Laura Versmee, Mikael Dusenne, Niloofar Jalali
}
