% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.name.R
\name{study.name}
\alias{study.name}
\title{Gets you the full name of the study on dbGaP}
\usage{
study.name(phs)
}
\arguments{
\item{phs}{dbGaP study ID (phs00xxxx, or 00xxxx, or xxx)}
}
\value{
Returns the full name of the study
}
\description{
Gets you the full name of the study on dbGaP
}
\author{
Gregoire Versmee, Laura Versmee, Mikael Dusenne, Niloofar Jalali
}
