% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub.study.R
\name{sub.study}
\alias{sub.study}
\title{Gives the substudies associated with the one selected}
\usage{
sub.study(phs)
}
\arguments{
\item{phs}{dbGap study ID (phs00xxxx, or 00xxxx, or xxx)}
}
\value{
Returns a dataframe with 2 cols : the phs id and the name of the substudy
}
\description{
Gives the substudies associated with the one selected
}
\author{
Gregoire Versmee, Laura Versmee
}
