% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_rep.R
\name{variables.report}
\alias{variables.report}
\title{Create a table 1 with the summary statistics of numerical variables of the study}
\usage{
variables.report(phs)
}
\arguments{
\item{phs}{dbGap study ID (phs00xxxx, or 00xxxx, or xxx)}
}
\value{
an htmlTable
}
\description{
This function extracts informations from var_report.xml files from the dbgap ftp server to create a variable dictionnary.
}
\author{
Gregoire Versmee, Laura Versmee
}
