% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.dbgap.R
\name{browse.dbgap}
\alias{browse.dbgap}
\title{Open the dbGap webpag of your study in your browser}
\usage{
browse.dbgap(phs, jupyter = FALSE)
}
\arguments{
\item{phs}{dbGap study ID (phs00xxxx, or 00xxxx, or xxx)}

\item{jupyter}{set on TRUE if you are in a jypyterhub environment}
}
\value{
Open the dbGap webpag of your study in your browser
}
\description{
Open the dbGap webpag of your study in your browser
}
\author{
Gregoire Versmee, Laura Versmee
}
