% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.study.R
\name{browse.study}
\alias{browse.study}
\title{Open the study webpage in a web browser}
\usage{
browse.study(phs, jupyter = FALSE)
}
\arguments{
\item{phs}{dbGap study ID (phs00xxxx, or 00xxxx, or xxx)}

\item{jupyter}{set on TRUE if you are in a jypyterhub environment}
}
\value{
Open the study webpage in a web browser
}
\description{
Open the study webpage in a web browser
}
\author{
Gregoire Versmee, Laura Versmee
}
