% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consent.groups.R
\name{consent.groups}
\alias{consent.groups}
\title{Gets you the list of the consent groups of the study}
\usage{
consent.groups(phs)
}
\arguments{
\item{phs}{dbGap study ID (phs00xxxx, or 00xxxx, or xxx)}
}
\value{
Returns a data frame with 3 cols : "consent group number", "cg short name" and "cg long name"
}
\description{
Gets you the list of the consent groups of the study
}
\author{
Gregoire Versmee, Laura Versmee
}
