% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables.dict.R
\name{datatables.dict}
\alias{datatables.dict}
\title{Create a tables dictionnary of your study}
\usage{
datatables.dict(phs)
}
\arguments{
\item{phs}{dbGap ID of your study "phs000000". No version please, as it will automatically return the latest one.}
}
\value{
a data.frame with 3 cols : XXXXX
}
\description{
This function extracts informations from data.dict.xml files from the dbgap ftp server to create a table dictionnary.
}
\author{
Gregoire Versmee, Laura Versmee
}
