% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbgap.download.R
\name{dbgap.download}
\alias{dbgap.download}
\title{Decrypt ncbi_enc files}
\usage{
dbgap.download(krt, key = FALSE)
}
\arguments{
\item{file}{file or folder where your encrypted files are located}
}
\value{
Decrypt the file(s) and replace it in the same folder
}
\description{
This function decrypts dbGap files (*ncbi_enc) using your personnal key. Be careful, it can replace the settings of your "vdb-info" file
}
\author{
Gregoire Versmee, Laura Versmee
}
