% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.dbgap.R
\name{search.dbgap}
\alias{search.dbgap}
\title{Open the results of your search on dbGap}
\usage{
search.dbgap(term, jupyter = FALSE)
}
\arguments{
\item{term}{Term of your search}

\item{jupyter}{set on TRUE if you are in a jypyterhub environment}
}
\value{
Open the results of your search on dbGap
}
\description{
Open the results of your search on dbGap
}
\author{
Gregoire Versmee, Laura Versmee
}
