% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.dict.R
\name{variables.dict}
\alias{variables.dict}
\title{Creates a variables dictionary of your study}
\usage{
variables.dict(phs)
}
\arguments{
\item{phs}{dbGaP study ID (phs00xxxx, or 00xxxx, or xxx)}
}
\value{
A data.frame with 4 columns : variable identifier (dbGaP), table name, variable name, variable description
}
\description{
This function extracts informations from data.dict.xml files from the dbGaP ftp server to create a variable dictionary.
}
\author{
Gregoire Versmee, Laura Versmee, Mikael Dusenne, Niloofar Jalali
}
