% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.dict.R
\name{variables.dict}
\alias{variables.dict}
\title{Create a variables dictionnary of your study}
\usage{
variables.dict(phs)
}
\arguments{
\item{phs}{dbGap study ID (phs00xxxx, or 00xxxx, or xxx)}
}
\value{
a data.frame with 4 cols : variable identifier (dbGap), table name, variable name, variable description
}
\description{
This function extracts informations from data.dict.xml files from the dbgap ftp server to create a variable dictionnary.
}
\author{
Gregoire Versmee, Laura Versmee
}
